feat: added PlayerRulesEnforcer

+++ engine/src/main/java/com/shreyashurakadli/engine/state/player/Player.kt
@@ -3,7 +3,7 @@ package com.shreyashurakadli.engine.state.player
 import com.shreyashurakadli.engine.state.piece.Piece
 
 internal data class Player(
-    val id: UByte,
+    val id: Int,
     val name: String,
     val status: PlayerStatus,
     val pieces: List<Piece>
